+++ Source/Rhetos.Configuration.Autofac/ApplicationInitialization.cs
@@ -32,7 +32,7 @@ namespace Rhetos.Deployment
     /// </summary>
     public static class ApplicationInitialization
     {
-        public static Type[] GetSortedInitializers(UnitOfWorkScope scope)
+        public static Type[] GetSortedInitializers(IUnitOfWorkScope scope)
         {
             // The plugins in the container are sorted by their dependencies defined in ExportMetadata attribute (static typed):
             var initializers = scope.Resolve<IPluginsContainer<IServerInitializer>>().GetPlugins();