+++ Source/Rhetos.Configuration.Autofac/Modules/DatabaseRuntimeModule.cs
@@ -18,7 +18,6 @@
 */
 
 using Autofac;
-using Rhetos.Persistence;
 using Rhetos.Utilities;
 
 namespace Rhetos.Configuration.Autofac.Modules